Fix pthread_setname_np race condition in thread_base::event_loop
The thread->m_thread member is set via pthread_create(3). However,
it is not guaranteed that pthread_create writes this member before
it returns. Since thread_base::event_loop is the function executed
by pthread_create, we cannot assume that it is set. Instead, use
pthread_self() which is guaranteed to be set.